search_service/requirements.txt
unstructured[all-docs]
pytesseract
pillow
llama-index
llama-index-llms-google-genai
llama-index-llms-bedrock-converse
llama-index-embeddings-google-genai
llama-index-vector-stores-milvus
llama-index-readers-elasticsearch
llama-index-vector-stores-opensearch
transformers
torch
torchaudio
python-dotenv
llama-index-embeddings-clip 
git+https://github.com/openai/CLIP.git 
torchvision
python-magic-bin
easyocr
watchdog
textdivider
fastapi
sentence-transformers
FlagEmbedding
google-genai
opencv-python